WIP: Implement Moving Distance to moving objects + command
